Refactor graphs.py into layered spacex_graphs package
Split the 752-line script into modules with single responsibilities
(config, cache, parsing, transform, plotting, output, cli) behind a
thin graphs.py shim so the CI/Docker entry point is unchanged.

- Remove all import-time side effects (network fetches, global state)
- Deduplicate Falcon/Starship date parsing into one parser
- Fetch each page once: 304 results drive the skip decision, halving
  requests to Wikipedia
- Skip logic now only applies to --output; display mode always renders
- Replace positional tuples with a LaunchRecord NamedTuple
- Replace deprecated BeautifulSoup findAll with find_all
- Add unittest suites for the pure parsing/transform functions
