+++ spacex_graphs/__init__.py
@@ -1,0 +1,1 @@
+"""Generates graphs of SpaceX mass-to-orbit launches from Wikipedia data."""
