+++ spacex_graphs/__main__.py
@@ -1,0 +1,6 @@
+"""Allows running the package directly: python3 -m spacex_graphs"""
+
+from spacex_graphs.cli import main
+
+if __name__ == "__main__":
+    main()
